# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vai Me Bancar API",
        "description": "API para sistema de vaquinhas com integração ao Asaas",
        "contact": {
            "email": "contato@vaimebancar.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://vaimebancar.codegus.com",
            "description": "Servidor de produção"
        }
    ],
    "paths": {
...
        "/api/projects": {
            "get": {
                "tags": [
                    "Projetos"
                ],
                "summary": "Listar todos os projetos",
                "operationId": "8d4a13ef3fc773ff4031e515460919ea",
                "responses": {
                    "200": {
                        "description": "Lista de projetos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Project"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Projetos"
                ],
                "summary": "Criar um novo projeto",
                "operationId": "a4dd4c17bf5fc24c224357f03dcc848a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "budget",
                                    "start_date",
                                    "end_date",
                                    "owner_name",
                                    "cellphone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Vaquinha do João"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "João precisa de dinheiro para comprar um novo celular"
                                    },
                                    "budget": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-20"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-02-20"
                                    },
                                    "owner_name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "cellphone": {
                                        "type": "string",
                                        "example": "11999999999"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projeto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Projeto criado com sucesso!"
                                        },
                                        "project": {
                                            "$ref": "#/components/schemas/Project"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/projects/{project}": {
            "get": {
                "tags": [
                    "Projetos"
                ],
                "summary": "Visualizar projeto específico",
                "operationId": "1352e06e55f6ab10d3702509cd385918",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "ID do projeto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projeto retornado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "project": {
                                            "$ref": "#/components/schemas/Project"
                                        },
                                        "fundraising_stats": {
                                            "properties": {
                                                "help_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 500
                                                },
                                                "stop_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 300
                                                },
                                                "total_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 800
                                                },
                                                "help_percentage": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 62.5
                                                },
                                                "stop_percentage": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 37.5
                                                },
                                                "stop_wins": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "troll_message": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "help_count": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "stop_count": {
                                                    "type": "integer",
                                                    "example": 3
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projeto não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Project": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Vaquinha do João"
                    },
                    "description": {
                        "type": "string",
                        "example": "João precisa de dinheiro para comprar um novo celular"
                    },
                    "budget": {
                        "type": "number",
                        "format": "float",
                        "example": 1500
                    },
                    "current_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-20T00:00:00.000000Z"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-20T00:00:00.000000Z"
                    },
                    "owner_name": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "cellphone": {
                        "type": "string",
                        "example": "11999999999"
                    },
                    "category": {
                        "type": "string",
                        "example": "Shark Tank"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Erro de validação"
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    },
...
}